devhunt/requirements.txt
anyio==4.2.0
apparmor==2.13.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Automat==0.8.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
boto3==1.34.58
botocore==1.34.58
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloud-init==23.4.4
colorama==0.4.3
comm==0.2.1
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.1.1
cryptography==42.0.2
cssselect==1.2.0
cycler==0.12.1
dbus-python==1.2.16
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.4.0
distro-info==0.23+ubuntu1.1
dos2unix==1
entrypoints==0.3
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.8.0
fastjsonschema==2.19.1
filelock==3.13.1
Flask==3.0.2
fonttools==4.48.1
fqdn==1.5.1
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.14.0
httpx==0.26.0
huggingface-hub==0.21.4
hyperlink==19.0.0
idna==2.8
importlib-metadata==7.0.1
importlib-resources==6.1.1
incremental==16.10.1
iotop==0.6
ipykernel==6.29.2
ipython==8.12.3
ipywidgets==8.1.2
isoduration==20.11.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
JCC==3.11
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpatch==1.22
jsonpointer==2.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.12
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
jupyterlab_widgets==3.0.10
keyring==18.0.1
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
LibAppArmor==2.13.3
lucene==8.11.0
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib==3.7.4
matplotlib-inline==0.1.6
mistune==3.0.2
more-itertools==4.2.0
mpmath==1.3.0
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.6.0
netifaces==0.10.4
networkx==3.1
notebook==7.0.8
notebook_shim==0.2.3
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandocfilters==1.5.1
parsel==1.8.1
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
pillow==10.2.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
praw==7.7.1
prawcore==2.4.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
Protego==0.3.0
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
PyDispatcher==2.0.7
pydos2unix==0.0.1
Pygments==2.17.2
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pylibacl==0.5.4
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==24.0.0
pyparsing==3.1.1
pyrsistent==0.15.5
pyserial==3.4
PySocks==1.7.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
pyxattr==0.6.1
PyYAML==5.3.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
queuelib==1.6.2
rdiff-backup==2.0.0
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
requests-unixsocket==0.2.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
s3transfer==0.10.0
sacremoses==0.1.1
safetensors==0.4.2
Scrapy==2.11.0
SecretStorage==2.3.1
selenium==4.17.2
Send2Trash==1.8.2
sentencepiece==0.2.0
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
sos==4.5.6
soupsieve==2.5
ssh-import-id==5.10
stack-data==0.6.3
sympy==1.12
systemd-python==234
terminado==0.18.0
tinycss2==1.2.1
tldextract==5.1.1
tokenizers==0.15.2
tomli==2.0.1
torch==2.2.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.2
trio==0.24.0
trio-websocket==0.11.1
triton==2.2.0
Twisted==18.9.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
update-checker==0.18.0
uri-template==1.3.0
urllib3==1.26.18
w3lib==2.1.2
wadllib==1.3.3
wcwidth==0.2.13
webcolors==1.13
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.10
wsproto==1.2.0
zipp==3.17.0
zope.interface==6.1
